Move the position package into the compiler package.
