+++ internal/vm/compiler/ast/ast.go
@@ -8,7 +8,7 @@ import (
 	"time"
 
 	"github.com/google/mtail/internal/metrics"
-	"github.com/google/mtail/internal/vm/position"
+	"github.com/google/mtail/internal/vm/compiler/position"
 	"github.com/google/mtail/internal/vm/symbol"
 	"github.com/google/mtail/internal/vm/types"
 )
