+++ internal/vm/compiler/ast/walk_test.go
@@ -9,7 +9,7 @@ import (
 	"github.com/google/mtail/internal/testutil"
 	"github.com/google/mtail/internal/vm/compiler/ast"
 	"github.com/google/mtail/internal/vm/compiler/parser"
-	"github.com/google/mtail/internal/vm/position"
+	"github.com/google/mtail/internal/vm/compiler/position"
 	"github.com/google/mtail/internal/vm/types"
 )
