+++ internal/vm/compiler/codegen/codegen.go
@@ -16,8 +16,8 @@ import (
 	"github.com/google/mtail/internal/vm/compiler/ast"
 	"github.com/google/mtail/internal/vm/compiler/errors"
 	"github.com/google/mtail/internal/vm/compiler/parser"
+	"github.com/google/mtail/internal/vm/compiler/position"
 	"github.com/google/mtail/internal/vm/object"
-	"github.com/google/mtail/internal/vm/position"
 	"github.com/google/mtail/internal/vm/symbol"
 	"github.com/google/mtail/internal/vm/types"
 )
