+++ internal/vm/compiler/position/position.go
