+++ internal/vm/symbol/symtab.go
@@ -7,7 +7,7 @@ import (
 	"bytes"
 	"fmt"
 
-	"github.com/google/mtail/internal/vm/position"
+	"github.com/google/mtail/internal/vm/compiler/position"
 	"github.com/google/mtail/internal/vm/types"
 )
